requirements.txt
ftfy
regex
matplotlib
numpy
Pillow
pycocotools
pycocoevalcap
scikit-learn
scikit-image
tqdm
wandb
lmdb
contexttimer
decord
einops>=0.4.1
fairscale==0.4.4
iopath
ipython
omegaconf
opencv-python-headless==4.5.5.64
opendatasets
packaging
pandas
plotly
pre-commit
python-magic
sentencepiece
spacy
streamlit
timm==0.4.12
transformers>=4.25.0,<4.27
webdataset
wheel
protobuf==3.20.0
torchmetrics
torch==1.13.1+cu117
torchvision==0.14.1+cu117
--extra-index-url https://download.pytorch.org/whl/cu117